ollama_app/requirements.txt
kafka-python
requests
redis